Move header_control from common to openedx/core
