+++ lms/envs/common.py
@@ -1088,7 +1088,7 @@ MIDDLEWARE_CLASSES = (
     'request_cache.middleware.RequestCache',
 
     'mobile_api.middleware.AppVersionUpgrade',
-    'header_control.middleware.HeaderControlMiddleware',
+    'openedx.core.djangoapps.header_control.middleware.HeaderControlMiddleware',
     'microsite_configuration.middleware.MicrositeMiddleware',
     'django_comment_client.middleware.AjaxExceptionMiddleware',
     'django.middleware.common.CommonMiddleware',
